# docs/Swagger/api-docs.json
{"openapi":"3.0.1","info":{"title":"Cinema API","description":"Dokumentation for RestfulKino API af codeQuest","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/showings/{id}":{"get":{"tags":["showing-controller"],"summary":"Get one showing by id","description":"Returns a showing by id","operationId":"getShowingById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"A showing with given id","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Showing"}}}},"400":{"description":"No showing found with given id"}}},"put":{"tags":["showing-controller"],"summary":"Update a showing","description":"Returns a showing after updating it in the database","operationId":"updateShowing","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Showing"}}},"required":true},"responses":{"200":{"description":"A showing has been updated","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Showing"}}}},"400":{"description":"Could not update showing"}}},"delete":{"tags":["showing-controller"],"summary":"Delete a showing","description":"Deletes a showing","operationId":"deleteShowing","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"A showing has been deleted"},"400":{"description":"Could not delete showing"}}}},"/movies/{id}":{"get":{"tags":["movie-controller"],"summary":"Get one movie by id","description":"Returns a movie by id","operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"A movie with given id","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Movie"}}}},"400":{"description":"No movie found with given id"}}},"put":{"tags":["movie-controller"],"summary":"Update a movie","description":"Returns a movie after updating","operationId":"updateMovie","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Movie"}}},"required":true},"responses":{"200":{"description":"A movie has been updated","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Movie"}}}},"400":{"description":"Could not update movie"}}}},"/halls/{id}":{"get":{"tags":["hall-controller"],"summary":"Get one hall by id","description":"Returns a hall by id","operationId":"getHallById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"A hall with given id","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Hall"}}}},"400":{"description":"No hall found with given id"}}},"put":{"tags":["hall-controller"],"summary":"Update a hall","description":"Returns a hall after updating it in the database","operationId":"updateHall","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Hall"}}},"required":true},"responses":{"200":{"description":"A hall has been updated","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Hall"}}}},"400":{"description":"Could not update hall"}}}},"/cinemas/{id}":{"get":{"tags":["cinema-controller"],"summary":"Get one cinema by id","description":"Returns a cinema by id","operationId":"getCinemaById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"A cinema with given id","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Cinema"}}}},"400":{"description":"No cinema found with given id"}}},"put":{"tags":["cinema-controller"],"summary":"Update a cinema","description":"Updates a cinema","operationId":"updateCinema","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Cinema"}}},"required":true},"responses":{"200":{"description":"A cinema has been deleted","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Cinema"}}}},"400":{"description":"Could not delete cinema"}}}},"/showings":{"get":{"tags":["showing-controller"],"summary":"Get all showings","description":"Returns a list of all showings","operationId":"getAllShowings","responses":{"200":{"description":"A list of all showings","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ShowingResponseDto"}}}}},"400":{"description":"No showings found"}}},"post":{"tags":["showing-controller"],"summary":"Add a showing","description":"Returns a showing after adding it to the database","operationId":"addShowing","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Showing"}}},"required":true},"responses":{"200":{"description":"A showing has been added","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Showing"}}}},"400":{"description":"Could not add showing"}}}},"/reservations":{"get":{"tags":["reservation-controller"],"summary":"Get all reservations","description":"Returns a list of all reservations","operationId":"getAllReservations","responses":{"200":{"description":"A list of all reservations","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ReservationResponseDto"}}}}},"400":{"description":"No reservations found"}}},"post":{"tags":["reservation-controller"],"summary":"Add a reservation","description":"Returns a reservation after adding","operationId":"addReservation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Reservation"}}},"required":true},"responses":{"200":{"description":"Reservation added","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Reservation"}}}},"400":{"description":"Reservation not added"}}}},"/movies/{imdbId}":{"post":{"tags":["movie-controller"],"summary":"Add a movie","description":"Adds a movie","operationId":"addMovie","parameters":[{"name":"imdbId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"A movie has been added","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Movie"}}}},"400":{"description":"Could not add movie"}}}},"/halls":{"get":{"tags":["hall-controller"],"summary":"Get all halls","description":"Returns a list of all halls","operationId":"getAllHalls","responses":{"200":{"description":"A list of all halls","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/HallResponseDto"}}}}},"400":{"description":"No halls found"}}},"post":{"tags":["hall-controller"],"summary":"Create a hall","description":"Returns a hall after adding it to the database","operationId":"createHall","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Hall"}}},"required":true},"responses":{"200":{"description":"A hall has been created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Hall"}}}},"400":{"description":"Could not create hall"}}}},"/cinemas":{"get":{"tags":["cinema-controller"],"summary":"Get all cinemas","description":"Returns a list of all cinemas","operationId":"getCinemas","responses":{"200":{"description":"A list of all cinemas","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Cinema"}}}}},"400":{"description":"No cinemas found"}}},"post":{"tags":["cinema-controller"],"summary":"Add a cinema","description":"Returns a cinema after adding it to the database","operationId":"addCinema","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Cinema"}}},"required":true},"responses":{"200":{"description":"A cinema has been added","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Cinema"}}}},"400":{"description":"Could not add cinema"}}}},"/api/user-with-role":{"post":{"tags":["user-with-role-controller"],"summary":"Add a new UserWithRoles user","description":"If a default role is defined (app.default-role ), this role will be assigned to the user.","operationId":"addUserWithRoles","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserWithRolesRequest"}}},"required":true},"responses":{"200":{"description":"A new user has been added","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserWithRolesResponse"}}}},"400":{"description":"Could not add user","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserWithRolesResponse"}}}}}}},"/api/auth/login":{"post":{"tags":["authentication-controller"],"summary":"Login","description":"Use this to login and get a token","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successfully logged in and got a token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"400":{"description":"Could not log in"}}}},"/api/user-with-role/remove-role/{username}/{role}":{"patch":{"tags":["user-with-role-controller"],"summary":"Removes a role from a UserWithRoles","description":"Caller must be authenticated with the role ADMIN","operationId":"removeRole","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string"}},{"name":"role","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"A role has been removed from user","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserWithRolesResponse"}}}},"400":{"description":"Could not remove role from user","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserWithRolesResponse"}}}}}}},"/api/user-with-role/add-role/{username}/{role}":{"patch":{"tags":["user-with-role-controller"],"summary":"Add a role to a UserWithRoles","description":"Caller must be authenticated with the role ADMIN","operationId":"addRole","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string"}},{"name":"role","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"A role has been added","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserWithRolesResponse"}}}},"400":{"description":"Could not add role to user","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserWithRolesResponse"}}}}}}},"/showings/{id}/takenSeats":{"get":{"tags":["showing-controller"],"summary":"Get unavailable seats","description":"Returns a list of all unavailable seats for a showing by id","operationId":"getUnavailableSeats","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"A list of all unavailable seats for a showing with given id","content":{"*/*":{"schema":{"type":"array","items":{"type":"integer","format":"int32"}}}}},"400":{"description":"No showing found with given id"}}}},"/showings/movie/{id}":{"get":{"tags":["showing-controller"],"summary":"Get all showings by movie","description":"Returns a list of all showings by movie","operationId":"getAllShowingsByMovie","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"A list of all showings by movie","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ShowingResponseDto"}}}}},"400":{"description":"No showings found by movie"}}}},"/reservations/{id}":{"get":{"tags":["reservation-controller"],"summary":"Get one reservation by id","description":"Returns a reservation by id","operationId":"getReservationByReservationId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"A reservation with given id","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Reservation"}}}},"400":{"description":"No reservation found with given id"}}},"delete":{"tags":["reservation-controller"],"summary":"Deletes a reservation","description":"Deletes a reservation with given id","operationId":"deleteReservation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Reservation deleted"},"400":{"description":"Reservation not deleted"}}}},"/reservations/user/{user}":{"get":{"tags":["reservation-controller"],"summary":"Get all reservations by username","description":"Returns a list of all reservations by username","operationId":"getReservationsByUserName","parameters":[{"name":"user","in":"path","required":true,"schema":{"$ref":"#/components/schemas/UserWithRoles"}}],"responses":{"200":{"description":"A list of all reservations with given username","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ReservationResponseDto"}}}}},"400":{"description":"No reservations found with given username"}}}},"/reservations/showing/{id}":{"get":{"tags":["reservation-controller"],"summary":"Get all reservations by showing id","description":"Returns a list of all reservations by showing id","operationId":"getReservationsByShowingId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"A list of all reservations with given showing id","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ReservationResponseDto"}}}}},"400":{"description":"No reservations found with given showing id"}}}},"/movies":{"get":{"tags":["movie-controller"],"summary":"Get all movies","description":"Returns a list of all movies","operationId":"getAll","responses":{"200":{"description":"A list of all movies","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MovieResponseDto"}}}}},"400":{"description":"No movies found"}}}},"/movies/imdbid/{imdbId}":{"get":{"tags":["movie-controller"],"summary":"Get one movie by imdbId","description":"Returns a movie by imdbId","operationId":"getMovie","parameters":[{"name":"imdbId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"A movie with given imdbId","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Movie"}}}},"400":{"description":"No movie found with given imdbId"}}}},"/halls/{id}/seats":{"get":{"tags":["hall-controller"],"summary":"Get all seats for a hall","description":"Returns a list of all seats for a hall by id","operationId":"getSeatsForHall","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"A list of all seats for a hall with given id","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SeatResponseDto"}}}}},"400":{"description":"No hall found with given id"}}}},"/halls/cinema/{id}":{"get":{"tags":["hall-controller"],"summary":"Get one hall by cinema id","description":"Returns a hall by cinema id","operationId":"getHallByCinemaId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"A hall with given cinema id","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Hall"}}}},"400":{"description":"No hall found with given cinema id"}}}},"/api/user-with-role/users":{"get":{"tags":["user-with-role-controller"],"summary":"Get all users with roles","description":"Returns a list of all users with their roles","operationId":"getAllUsersWithRoles","responses":{"200":{"description":"A list of all users with their roles","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserWithRolesResponse"}}}}},"400":{"description":"No users found","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserWithRolesResponse"}}}}}}}}},"components":{"schemas":{"Cinema":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"address":{"type":"string"}}},"Hall":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"cinema":{"$ref":"#/components/schemas/Cinema"},"hallNumber":{"type":"integer","format":"int32"},"supports3d":{"type":"boolean"},"supportsImax":{"type":"boolean"},"seatRows":{"type":"integer","format":"int32"},"seatsPerRow":{"type":"integer","format":"int32"},"totalSeats":{"type":"integer","format":"int32"}}},"Movie":{"type":"object","properties":{"title":{"type":"string"},"id":{"type":"integer","format":"int64"},"year":{"type":"string"},"rated":{"type":"string"},"released":{"type":"string"},"runtime":{"type":"string"},"genre":{"type":"string"},"director":{"type":"string"},"writer":{"type":"string"},"actors":{"type":"string"},"plot":{"type":"string"},"plotDK":{"type":"string"},"poster":{"type":"string"},"metascore":{"type":"string"},"imdbRating":{"type":"string"},"imdbVotes":{"type":"string"},"imdbID":{"type":"string"},"website":{"type":"string"},"response":{"type":"string"},"supports3d":{"type":"boolean"}}},"Showing":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"movie":{"$ref":"#/components/schemas/Movie"},"hall":{"$ref":"#/components/schemas/Hall"},"startTime":{"type":"string","format":"date-time"},"is3d":{"type":"boolean"},"isImax":{"type":"boolean"}}},"GrantedAuthority":{"type":"object","properties":{"authority":{"type":"string"}}},"PriceClass":{"type":"object","properties":{"name":{"type":"string"},"price":{"type":"number","format":"double"}}},"Reservation":{"required":["price","reservedSeats","showing"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"showing":{"$ref":"#/components/schemas/Showing"},"price":{"type":"number","format":"double"},"user":{"$ref":"#/components/schemas/UserWithRoles"},"email":{"type":"string"},"reservedSeats":{"type":"array","items":{"$ref":"#/components/schemas/Seat"}}}},"Role":{"type":"object","properties":{"roleName":{"type":"string"}}},"Seat":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"hall":{"$ref":"#/components/schemas/Hall"},"seatNumber":{"type":"integer","format":"int32"},"seatRowNumber":{"type":"integer","format":"int32"},"priceClass":{"$ref":"#/components/schemas/PriceClass"}}},"UserWithRoles":{"type":"object","properties":{"username":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"enabled":{"type":"boolean"},"created":{"type":"string","format":"date-time"},"edited":{"type":"string","format":"date-time"},"roles":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/Role"}},"credentialsNonExpired":{"type":"boolean"},"authorities":{"type":"array","items":{"$ref":"#/components/schemas/GrantedAuthority"}},"accountNonExpired":{"type":"boolean"},"accountNonLocked":{"type":"boolean"}}},"UserWithRolesRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"email":{"type":"string"}}},"UserWithRolesResponse":{"type":"object","properties":{"userName":{"type":"string"},"roleNames":{"type":"array","items":{"type":"string"}},"email":{"type":"string"}}},"LoginRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"LoginResponse":{"type":"object","properties":{"username":{"type":"string"},"token":{"type":"string"},"roles":{"type":"array","items":{"type":"string"}}}},"HallResponseDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"hallNumber":{"type":"integer","format":"int32"},"cinema":{"$ref":"#/components/schemas/Cinema"}}},"MovieResponseDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string"},"imdbRating":{"type":"string"},"poster":{"type":"string"},"genre":{"type":"string"},"year":{"type":"string"}}},"ShowingResponseDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"movie":{"$ref":"#/components/schemas/MovieResponseDto"},"hall":{"$ref":"#/components/schemas/HallResponseDto"},"startTime":{"type":"string","format":"date-time"},"is3d":{"type":"boolean"},"isImax":{"type":"boolean"}}},"ReservationResponseDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"user":{"$ref":"#/components/schemas/UserWithRolesResponse"},"showing":{"$ref":"#/components/schemas/ShowingResponseDto"},"price":{"type":"number","format":"double"},"reservedSeats":{"type":"array","items":{"$ref":"#/components/schemas/SeatResponseDto"}}}},"SeatResponseDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"seatNumber":{"type":"integer","format":"int32"},"seatRowNumber":{"type":"integer","format":"int32"},"price":{"type":"number","format":"double"},"priceClass":{"type":"string"}}}}}}